requirements.txt
pyserial
pyubx2